feat new desing for landing page
